Transforming some integers to bigint.

+++ db/migrate/20190611094224_create_pregnant_statuses.rb
@@ -1,8 +1,8 @@
 class CreatePregnantStatuses < ActiveRecord::Migration[5.2]
   def change
     create_table :pregnant_statuses, :primary_key => :pregnant_status_id do |t|
-     	t.integer    :concept_id
-      t.integer      :encounter_id
+     	t.bigint    :concept_id
+      t.bigint      :encounter_id
     	t.integer    :value_coded
       t.boolean    :voided, null: false, default: 0
       t.bigint        :voided_by
