+++ db/migrate/20190611095244_create_lab_orders.rb
@@ -3,7 +3,7 @@ class CreateLabOrders < ActiveRecord::Migration[5.2]
     create_table    :lab_orders , :primary_key => :lab_order_id do |t|
     	t.string         :tracking_number
     	t.datetime     :order_date
-    	t.integer       :encounter_id    	
+    	t.bigint       :encounter_id
     	t.bigint         :creator
     	t.boolean     :voided, null: false, default: 0
     	t.bigint        :voided_by
