+++ db/migrate/20190611095700_create_occupations.rb
@@ -1,7 +1,7 @@
 class CreateOccupations < ActiveRecord::Migration[5.2]
   def change
     create_table     :occupations , :primary_key => :occupation_id do |t|
-    	t.integer      :person_id, null: false
+    	t.bigint      :person_id, null: false
     	t.integer      :occupation, null: false
       t.bigint        :creator
     	t.boolean      :voided, null: false, default: 0
